+++ src/main/java/a_iterator/CustomIteratorImplementation.java
@@ -4,9 +4,7 @@
  * @see a_iterator.CustomListIterator
  * @since 11/01/2023
  */
-
-
-package iterator;
+package a_iterator;
 
 import java.util.ArrayList;
 import java.util.List;
