+++ src/main/java/a_iterator/GetAndLoopIterator.java
@@ -3,9 +3,7 @@
  * @author David Kariuki
  * @since 11/01/2023
  */
-
-
-package iterator;
+package a_iterator;
 
 import java.util.*;
