+++ src/main/java/b_iterable/IterateViaForEachLoop.java
@@ -1,11 +1,10 @@
 /**
- * @note This example first creates a new List and adds 3 elements to it.
- * Then it uses a for-each loop to iterate the elements of the List, and print out the toString() value of each element.
+ * @note This example first creates a new List and adds 3 elements to it. Then it uses a for-each
+ *     loop to iterate the elements of the List, and print out the toString() value of each element.
  * @author David Kariuki
  * @since 11/01/2023
  */
-
-package iterable;
+package b_iterable;
 
 import java.util.ArrayList;
 import java.util.List;
