+++ src/main/java/c_collection/b_java_collection_methods/Sort.java
@@ -19,7 +19,7 @@ public class Sort {
    */
   public static void main(String[] args) {
 
-    List<String> list = new ArrayList<String>();
+    List<String> list = new ArrayList<>();
     list.add("one");
     list.add("two");
     list.add("three");
